group and group nav rename from plan to group
