+++ client/src/modules/group/group-nav/group-nav.css
@@ -1,4 +1,4 @@
-.plan-nav-container {
+.group-nav-container {
   display: flex;
   order: 99;
   flex: 0 0 auto;
@@ -9,7 +9,7 @@
   border-top: 1px solid var(--da);
 }
 
-.plan-nav-item-container {
+.group-nav-item-container {
   flex: 1 0 0;
   display: flex;
   height: 64px;
@@ -17,11 +17,11 @@
   border-right: 1px solid var(--la);
 }
 
-.plan-nav-item-container:last-child {
+.group-nav-item-container:last-child {
   border-right: none;
 }
 
-.plan-nav-button {
+.group-nav-button {
   width: 100%;
   height: 100%;
   padding: 5px;
@@ -33,25 +33,25 @@
   text-decoration: none !important;
 }
 
-.plan-nav-button span {
+.group-nav-button span {
   font-size: .7em;
   font-variant: small-caps;
 }
 
-.plan-nav-button .material-icons,
-.plan-nav-button span {
+.group-nav-button .material-icons,
+.group-nav-button span {
   color: var(--dk5);
 }
 
-.plan-nav-active {
+.group-nav-active {
   background: var(--m5);
 }
 
-.plan-nav-active .material-icons,
-.plan-nav-active .plan-nav-button span {
+.group-nav-active .material-icons,
+.group-nav-active .group-nav-button span {
   color: white;
 }
 
-.plan-nav-item-container:hover:not(.plan-nav-active) {
+.group-nav-item-container:hover:not(.group-nav-active) {
   background-color: var(--m2);
 }
