+++ client/src/modules/group/group.css
@@ -1,4 +1,4 @@
-.plan-container {
+.group-container {
   display: flex;
   flex-direction: column;
 }
