+++ client/src/modules/group/group.html
@@ -1,4 +1,4 @@
-<div class="plan-container h100 flex-col">
+<div class="group-container h100 flex-col">
   <ui-view class="flex-window"></ui-view>
-  <plan-nav class="flex-0"></plan-nav>
+  <group-nav class="flex-0"></group-nav>
 </div>
